Change collision checking and reaction
In both the X and Y axis, collision reaction seemed to not be putting
the player flush against tiles. So I switched it to instead move the
collision rect one axis at a time. I also realized doing it this way,
I didn't need to keep track of what tile the collision happened. And
I figured out that move_ip() was working as intended in my previous
way of doing collision reaction, I just wasn't thinking of the math
properly. The reason why -x_velocity or -y_velocity was the correct
value to give regardless of direction was because if, say, the player
is moving left (a negative direction), if I make a negative, negative,
it will be a positive (or right) push. And it worked in every direction.
But now I'm moving and checking for collisions 1 pixel at a time, so
having that nice math was no longer possible.

Unfortunately, Y Axis collision reaction still seems to be incorrect.
The player will float about 2 pixels above the ground after jumping.
Luckily, X Axis collision reaction seems to be working perfectly,
as the player goes flush up against a tile.