Fix collision reaction on Y not being flush
That in itself was an easy fix in code, but makes me wonder if my
collision rect offset isn't correct on the Y by one pixel. Regardless,
collision reaction is not quite perfected. The Player now will sometimes
be inside a tile after jumping and landing at an angle (literally at
the corner of their collision rect).

There is also a problem of unintentional variable jump height.