Fix collision reaction(!!!!)
The real problem was found! So the way that it checked the tiles around
the player was from perfect center left, right, top, and bottom. It
wasn't taking into consideration any area past or before those points.
So the reason why the player would sink in at corners was because when
coming in at an angle, neither center left/right or center top/bottom
noticed a collision even though there was one on that side. So I
realized I needed to take into consideration that ENTIRE side, not just
the center of each. And presto!! Proper working collision!!

I'm VERY happy :)

Big thanks to Jonathan Whiting's blog post on 2D Tilemap Collision!