Add world pos rect
When the camera shifts the game level or "world", that is not noticed by
any entity's pos_rect because the pos rect only exists in screen space.
So I need a way to track "world" space positioning by making it the
amount the camera shifts the game level and all its entities.

Since the camera shifts in the direction opposite of the way the target
is heading, increase the entities' world position by the same direction
the target is heading. This way when checking for where an entity is
for tile collision, it's based on where it would be in the world, not
just screen space.