Fix Actor collision checking to work via world_pos
pos_rect only exists in Screen Space, so the collision checking for
actors needs to be with world_pos (aka World Space). But the pos_rect
needs to still be updated so if the actor is the target focus,
the target will stay in focus (center of screen) AND if they jump, it
shows them going up and falling down.