Make shift_amount based on target's world_pos
This made fixing the collision so hard to track down and fix. At first,
the player wouldn't move, and that's because I was still shifting based
on pos_rect (aka ScreenSpace). I finally fixed that to be world_pos, but
then the player's (which is the camera target) world_pos was always
further out than they should have been visually. And it finally dawned
on me that I had the player's world_pos being moved via the collision
checking AND the camera. Removed that line and BAM, working world
collision! So I guess the lessoned learned was that the camera should
only shift the world and its entities, don't set the target's world pos
based on the camera!!