Remove reposting of TIMER_ENDED event
I thought I should be doing my due diligence to repost these events
so the right object would hear it, but this idea seems to be flawed.
When I repost an event, I don't repost it with the id of the Timer that
sent it, which causes a crash. Also, it would appear that is unessential
now, as both the Player and Game don't seem to read an event the other
was hoping to see (and apparently removing it from the queue by reading
it)